Moved XML tests to separate file
