+++ MarkMpn.Sql4Cds.Engine.Tests/MarkMpn.Sql4Cds.Engine.Tests.csproj
@@ -113,6 +113,7 @@
     <Compile Include="StubTableSizeCache.cs" />
     <Compile Include="WindowFunctionTests.cs" />
     <Compile Include="TempTableTests.cs" />
+    <Compile Include="XmlTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
